agregue promedio haciendo consulta sql de AVG, salio todo bien, se actualziza con js cada vez que se inserta o borra comentario

+++ assets/js/vue.js
@@ -4,19 +4,28 @@ document.addEventListener('DOMContentLoaded', async () => {
     let app = new Vue ({
         el: '#app',
         data: {
+            promedio: null,
             comentarios : null,
-            admin : false
+            admin : true
         }
     })
+    let idEvento = document.getElementById("evento").dataset.id;
 
+    GetPromedioPuntaje();
     GetComentarios();
-    
+
+    async function GetPromedioPuntaje() {
+        let r = await fetch("api/comentarios/promedio/"+idEvento)
+        let json = await r.json();
+        app.promedio = await json.promedio;
+    }
+
     async function GetComentarios() {
-        let idEvento = document.getElementById("evento").dataset.id;
         let r = await fetch("api/comentarios/eventos/"+idEvento)
         let json = await r.json();
         app.comentarios = await json;
         setTimeout(AsignarEventListener,0);
+        setTimeout(GetPromedioPuntaje,0);
         // catch(error => console.log(error));
     }
     
@@ -30,20 +39,20 @@ document.addEventListener('DOMContentLoaded', async () => {
 
     async function BorrarComentario(e){
         e.preventDefault();
-        let id = this.id;
-        let r = await fetch("api/comentarios/"+id,
-            {
-                "method":'DELETE'
-            });
-        console.log(r);
-        setTimeout(GetComentarios,0);
-        // catch(error => console.log(error));
+        if (app.admin){
+            let id = this.id;
+            let r = await fetch("api/comentarios/"+id,
+                {
+                    "method":'DELETE'
+                });
+            setTimeout(GetComentarios,0);
+            // catch(error => console.log(error));
+        }
     }
 
     async function EnviarComentario(e){
         e.preventDefault();
         let idUsuario = 1; // HARDCODEADO
-        let idEvento = document.getElementById("evento").dataset.id;
         let comentario = document.getElementById("comentario").value;
         let puntaje = document.getElementById("puntaje").value;
         let comentarioNuevo = {
@@ -57,7 +66,6 @@ document.addEventListener('DOMContentLoaded', async () => {
                 "method": "POST",
                 "body": JSON.stringify(comentarioNuevo)
         })
-        console.log(r);
         setTimeout(GetComentarios,0);
         // catch(error => console.log(error));
